+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - BREAKING: `Dir` now uses `inode()` and `inode_mut()` instead of `AsRef` and `AsMut` for consistent API with `File`.
+- `Dir::open` (deprecated for now) has become `Dir::open_inode`
 
 ## 0.1.0-alpha.3